% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{create_payload}
\alias{create_payload}
\title{Create payload for post request body}
\usage{
create_payload(id, format = "csv", labs = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Create payload for post request body
}
