% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{create_payload_v2}
\alias{create_payload_v2}
\title{Create payload for updated api: no id in the payload}
\usage{
create_payload_v2(format = "csv", labs = FALSE, ...)
}
\value{

}
\description{
Create payload for updated api: no id in the payload
}
