% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get-responses.R
\name{get_responses}
\alias{get_responses}
\title{Get an export of responses from a survey}
\usage{
get_responses(id, folder = tempdir(), fname = "qxre.zip", format = "spss",
  labs = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Get an export of responses from a survey
}
