% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get-responses.R
\name{get_responses_v2}
\alias{get_responses_v2}
\title{Get responses using current API version}
\usage{
get_responses_v2(
  ...,
  format = "spss",
  as_factor = TRUE,
  clean_names = TRUE,
  browse = FALSE
)
}
\arguments{
\item{browse}{}

\item{id}{Survey id as string or index from most recent list_surveys table}
}
\value{

}
\description{
Get responses using current API version
}
