% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_list-surveys.R
\name{list_surveys}
\alias{list_surveys}
\title{List surveys on the license, arrange by search term}
\usage{
list_surveys(search = "")
}
\value{

}
\description{
List surveys on the license, arrange by search term
}
