% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{list_to_json}
\alias{list_to_json}
\title{Convert R list into payload-usable json}
\usage{
list_to_json(d)
}
\arguments{
\item{d}{is a list of parameters and their values}
}
\description{
Convert R list into payload-usable json
}
