% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_share-surveys.R
\name{list_user_groups}
\alias{list_user_groups}
\title{List User Groups visible by current user (must be brand admin)}
\usage{
list_user_groups()
}
\description{
List User Groups visible by current user (must be brand admin)
}
