% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_print-survey.R
\name{print_survey}
\alias{print_survey}
\title{Convert survey list object to markdown and compile as pdf}
\usage{
print_survey(
  srv,
  file_path,
  keep_tex = TRUE,
  print_internal = FALSE,
  browse = FALSE
)
}
\arguments{
\item{browse}{}
}
\value{

}
\description{
Convert survey list object to markdown and compile as pdf
}
