% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_q-bar.R
\name{q_bar}
\alias{q_bar}
\title{Labelled bar chart uses qualtrics data}
\usage{
q_bar(dat, x, title_wrap = 40, col_wrap = 12)
}
\arguments{
\item{col_wrap}{}
}
\value{

}
\description{
Labelled bar chart uses qualtrics data
}
