% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{qp_html_to_tex}
\alias{qp_html_to_tex}
\title{Simple scan for html formatting tags -> latex}
\usage{
qp_html_to_tex(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Simple scan for html formatting tags -> latex
}
