% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{register_options}
\alias{register_options}
\title{register API KEY and root URL in env variables}
\usage{
register_options(conf = decrypt_conf())
}
\value{

}
\description{
register API KEY and root URL in env variables
}
