% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_scr-num.R
\name{score_scale}
\alias{score_scale}
\title{Generate a scale score for a set of columns}
\usage{
score_scale(srv, var_name, .vars, .rev = vars(matches("_r")), scale = "agree")
}
\arguments{
\item{scale}{}
}
\value{

}
\description{
Generate a scale score for a set of columns
}
