% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_scr-num.R
\name{scr_num}
\alias{scr_num}
\title{Recode scale with numeric values}
\usage{
scr_num(dat, .vars, .rev = NULL, scale = "agree", fct = FALSE)
}
\arguments{
\item{fct}{}
}
\description{
Recode scale with numeric values
}
