% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_share-surveys.R
\name{share_survey}
\alias{share_survey}
\title{Share a Survey with a User or Group}
\usage{
share_survey(survey_id, recipient_id, custom_permissions = list())
}
\arguments{
\item{survey_id}{is character: the Qx SurveyId to share}

\item{recipient_id}{is character: the Qx UserId or GroupId to share it with
#' @param custom_permissions
is a list that conforms to the following
scheme:
list(
surveyDefinitionManipulation = list(
copySurveyQuestions   = TRUE,
editSurveyFlow        = TRUE,
useBlocks             = TRUE,
useSkipLogic          = TRUE,
useConjoint           = TRUE,
useTriggers           = TRUE,
useQuotas             = TRUE,
setSurveyOptions      = TRUE,
editQuestions         = TRUE,
deleteSurveyQuestions = TRUE,
useTableOfContents    = TRUE,
useAdvancedQuotas     = TRUE
),
surveyManagement = list(
editSurveys           = TRUE,
activateSurveys       = TRUE,
deactivateSurveys     = TRUE,
copySurveys           = TRUE,
distributeSurveys     = TRUE,
deleteSurveys         = TRUE,
translateSurveys      = TRUE
),
response = list(
editSurveyResponses   = TRUE,
createResponseSets    = TRUE,
viewResponseId        = TRUE,
useCrossTabs          = TRUE,
useScreenouts         = TRUE
),
result = list(
downloadSurveyResults = TRUE,
viewSurveyResults     = TRUE,
filterSurveyResults   = TRUE,
viewPersonalData      = TRUE
)
)}
}
\description{
Share a Survey with a User or Group
}
