% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_utils.R
\name{encrypt_conf}
\alias{encrypt_conf}
\title{Encrypt qualtrics configuration details in project root}
\usage{
encrypt_conf(api_token, base_url = "ca1.qualtrics.com", key)
}
\arguments{
\item{key}{}
}
\description{
Encrypt qualtrics configuration details in project root
}
