% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_get-responses.R
\name{get_responses}
\alias{get_responses}
\title{Get an export of responses from a survey}
\usage{
get_responses(
  ...,
  folder = tempdir(),
  fname = "qxre.zip",
  format = "spss",
  labs = TRUE,
  as_factor = TRUE,
  clean_names = TRUE
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Get an export of responses from a survey
}
