% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_get-survey.R
\name{print_survey}
\alias{print_survey}
\title{Convert survey list object to markdown and compile as pdf}
\usage{
print_survey(srv, file_name, browse = FALSE)
}
\arguments{
\item{browse}{}
}
\description{
Convert survey list object to markdown and compile as pdf
}
