% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_q-tab.R
\name{qt_raw}
\alias{qt_raw}
\title{Labelled table for qualtrics scales}
\usage{
qt_raw(dat, .vars, .grp = groups(dat), title = NULL, browse = FALSE)
}
\arguments{
\item{browse}{}
}
\description{
Labelled table for qualtrics scales
}
